+++ src/commands/app/download.rs
@@ -23,6 +23,7 @@ use sprinkles::{
 
 use crate::{
     abandon,
+    handlers::{AppsDecider, ListApps},
     models::status::Info,
     output::colours::{bright_red, eprintln_yellow},
 };
@@ -34,7 +35,7 @@ pub struct Args {
     no_hash_check: bool,
 
     #[clap(help = "The packages to download")]
-    packages: Vec<package::Reference>,
+    apps: Vec<package::Reference>,
 
     #[clap(long, help = "Download new versions of all outdated apps")]
     outdated: bool,
@@ -47,14 +48,10 @@ impl super::Command for Args {
     const BETA: bool = true;
 
     async fn runner(self, ctx: &impl ScoopContext) -> Result<(), anyhow::Error> {
-        let packages = if self.packages.is_empty() {
-            if self.outdated {
-                list_outdated(ctx)?
-            } else {
-                abandon!("No packages provided")
-            }
-        } else {
-            self.packages
+        let packages = match AppsDecider::new(ctx, self.list_apps(), self.apps).decide()? {
+            Some(apps) if apps.is_empty() => abandon!("No apps selected"),
+            None => abandon!("No apps selected"),
+            Some(apps) => apps,
         };
 
         if self.no_hash_check {
@@ -133,10 +130,9 @@ impl super::Command for Args {
                 } else {
                     eprintln!();
                     let hash = result.computed_hash.no_prefix();
-                    pb.println(
-                        bright_red!("🔓 Hash mismatch: expected {actual_hash}, found {hash}",)
-                            .to_string(),
-                    );
+                    pb.println(bright_red!(
+                        "🔓 Hash mismatch: expected {actual_hash}, found {hash}",
+                    ));
                 }
             }
         }
@@ -145,34 +141,44 @@ impl super::Command for Args {
     }
 }
 
-fn list_outdated(ctx: &impl ScoopContext) -> Result<Vec<package::Reference>, anyhow::Error> {
-    let apps = install::Manifest::list_all_unchecked(ctx)?;
+impl Args {
+    fn list_apps<C: ScoopContext>(&self) -> impl ListApps<C> + use<C> {
+        let outdated = self.outdated;
+        move |ctx: &C| {
+            if outdated {
+                let apps = install::Manifest::list_all_unchecked(ctx)?;
 
-    Ok(apps
-        .par_iter()
-        .flat_map(|app| -> anyhow::Result<Info> {
-            if let Some(bucket) = &app.bucket {
-                let local_manifest = app.get_manifest(ctx)?;
-                // TODO: Add the option to check all buckets and find the highest version (will require semver to order versions)
-                let bucket = Bucket::from_name(ctx, bucket)?;
+                Ok(Some(
+                    apps.par_iter()
+                        .flat_map(|app| -> anyhow::Result<Info> {
+                            if let Some(bucket) = &app.bucket {
+                                let local_manifest = app.get_manifest(ctx)?;
+                                // TODO: Add the option to check all buckets and find the highest version (will require semver to order versions)
+                                let bucket = Bucket::from_name(ctx, bucket)?;
 
-                match Info::from_manifests(ctx, &local_manifest, &bucket) {
-                    Ok(info) => Ok(info),
-                    Err(err) => {
-                        error!(
-                            "Failed to get status for {}: {:?}",
-                            unsafe { app.name() },
-                            err
-                        );
-                        Err(err)?
-                    }
-                }
+                                match Info::from_manifests(ctx, &local_manifest, &bucket) {
+                                    Ok(info) => Ok(info),
+                                    Err(err) => {
+                                        error!(
+                                            "Failed to get status for {}: {:?}",
+                                            unsafe { app.name() },
+                                            err
+                                        );
+                                        Err(err)?
+                                    }
+                                }
+                            } else {
+                                error!("no bucket specified");
+                                anyhow::bail!("no bucket specified")
+                            }
+                        })
+                        .filter(|app| app.current != app.available)
+                        .map(|app| manifest::Reference::Name(app.name).into_package_ref())
+                        .collect::<Vec<_>>(),
+                ))
             } else {
-                error!("no bucket specified");
-                anyhow::bail!("no bucket specified")
+                anyhow::Ok(None)
             }
-        })
-        .filter(|app| app.current != app.available)
-        .map(|app| manifest::Reference::Name(app.name).into_package_ref())
-        .collect::<Vec<_>>())
+        }
+    }
 }