+++ src/commands/status.rs
@@ -259,33 +259,9 @@ impl Args {
                 .map(serde_json::to_value)
                 .collect::<Result<Vec<_>, _>>()?;
 
-            // if self.json {
-            //     let output = serde_json::to_string_pretty(&values)?;
-
-            //     println!("{output}");
-            // } else {
-            // TODO: Add a better way to add colours than this
-            // TODO: p.s this doesnt work atm
-            // use owo_colors::OwoColorize;
-            // let values = values
-            //     .into_par_iter()
-            //     .map(|mut value| {
-            //         if let Some(current) = value.get_mut("Current") {
-            //             *current = current.as_str().unwrap().red().to_string().into();
-            //         }
-
-            //         if let Some(available) = value.get_mut("Available") {
-            //             *available = available.as_str().unwrap().green().to_string().into();
-            //         }
-
-            //         value
-            //     })
-            //     .collect::<Vec<_>>();
-
             let outputs = Structured::new(&values).with_max_length(30);
 
             write!(output, "{outputs}")?;
-            // }
         }
 
         Ok(())