+++ src/output/structured.rs
@@ -109,7 +109,11 @@ impl Display for Structured {
                             contestants.push(base[i]);
                             contestants.push(element.len() + WALL.len());
 
-                            contestants.into_iter().max().unwrap()
+                            // Safe as we have just pushed at least two items into the vector in the lines above
+                            contestants
+                                .into_iter()
+                                .max()
+                                .expect("minimum 2 contestants")
                         })
                         .collect()
                 });