+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `app cleanup` command for removing old versions and cache entries
+
 ## [1.16.0] - 2025-19-01
 
 ### Fixes
@@ -18,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added config validations.
   - sfsu will now crash with an error message if `no_junction` is enabled.
 - Added `app download --outdated` flag to download new versions of all outdated apps
+- Returned logs in debug assertions going into current directory
 - Warnings in search command for deprecated usage
 - Support `json` flag in `search` command
 - Warning to help message for `json` flag calling out that it only works for certain commands
@@ -28,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed `json` flag from `app download` command
 - Download progress bars now show app name instead of url leaf
 - Download hash checks now report to a progress bar rather than a print message for each
+- Renamed `packages` parameter to `apps` in `app download` command (this should not affect usage at all)
 - Logs will now go into `<PWD>/logs` if running with debug assertions
 - `search` command no longer hides `[installed]` label if only searching for installed apps
 - Removed `disable_git` flag from `bucket add` command